// src/app/models/address.ts
export interface Address {
    id: number;
    userId: number;
    name: string;
    firstName: string;
    lastName: string;
    phone: string;
    street: string;
    streetNumber: number;
    department: string;
    departmentFloor: number;
    departmentTower: string;
    zipCode: string;
    state: string;
    city: string; 
}